Move short_term_storage related classes to managers
The ShortTermStorageManager is a manager and makes sense to have it live under galaxy.managers instead of galaxy.web.

It will also help reduce the dependency of the manager modules on web related stuff.
