+++ lib/galaxy/managers/markdown_util.py
@@ -46,6 +46,10 @@ from galaxy.managers.jobs import (
     summarize_job_parameters,
 )
 from galaxy.managers.licenses import LicensesManager
+from galaxy.managers.short_term_storage import (
+    ShortTermStorageMonitor,
+    storage_context,
+)
 from galaxy.model.item_attrs import get_item_annotation_str
 from galaxy.model.orm.now import now
 from galaxy.schema import PdfDocumentType
@@ -53,10 +57,6 @@ from galaxy.schema.tasks import GeneratePdfDownload
 from galaxy.util.markdown import literal_via_fence
 from galaxy.util.resources import resource_string
 from galaxy.util.sanitize_html import sanitize_html
-from galaxy.web.short_term_storage import (
-    ShortTermStorageMonitor,
-    storage_context,
-)
 from .markdown_parse import (
     GALAXY_MARKDOWN_FUNCTION_CALL_LINE,
     validate_galaxy_markdown,
