+++ lib/galaxy/managers/model_stores.py
@@ -9,6 +9,10 @@ from galaxy.jobs.manager import JobManager
 from galaxy.managers.context import ProvidesUserContext
 from galaxy.managers.export_tracker import StoreExportTracker
 from galaxy.managers.histories import HistoryManager
+from galaxy.managers.short_term_storage import (
+    ShortTermStorageMonitor,
+    storage_context,
+)
 from galaxy.managers.users import UserManager
 from galaxy.model.base import transaction
 from galaxy.model.scoped_session import galaxy_scoped_session
@@ -41,10 +45,6 @@ from galaxy.schema.tasks import (
 )
 from galaxy.structured_app import MinimalManagerApp
 from galaxy.version import VERSION
-from galaxy.web.short_term_storage import (
-    ShortTermStorageMonitor,
-    storage_context,
-)
 
 
 class ModelStoreUserContext(ProvidesUserContext):
