+++ lib/galaxy/managers/short_term_storage.py
