+++ lib/galaxy/webapps/galaxy/services/base.py
@@ -23,6 +23,10 @@ from galaxy.managers.base import (
 )
 from galaxy.managers.context import ProvidesUserContext
 from galaxy.managers.model_stores import create_objects_from_store
+from galaxy.managers.short_term_storage import (
+    ShortTermStorageAllocator,
+    ShortTermStorageTarget,
+)
 from galaxy.model import User
 from galaxy.model.store import (
     get_export_store_factory,
@@ -32,10 +36,6 @@ from galaxy.schema.fields import EncodedDatabaseIdField
 from galaxy.schema.schema import AsyncTaskResultSummary
 from galaxy.security.idencoding import IdEncodingHelper
 from galaxy.util import ready_name_for_url
-from galaxy.web.short_term_storage import (
-    ShortTermStorageAllocator,
-    ShortTermStorageTarget,
-)
 
 
 def ensure_celery_tasks_enabled(config):
