+++ test/unit/webapps/test_service_base.py
@@ -2,8 +2,8 @@ from typing import Tuple
 
 import pytest
 
+from galaxy.managers.short_term_storage import ShortTermStorageAllocator
 from galaxy.schema.schema import ModelStoreFormat
-from galaxy.web.short_term_storage import ShortTermStorageAllocator
 from galaxy.webapps.galaxy.services.base import model_store_storage_target
